// FacetsUI/src/app/modules/visitor/models/visitor-search.model.ts
export class VisitorSearchModel {
        isAssocifyMember: boolean;
        isRegisteredToFacets: boolean;
        firstName: string;
        lastName: string;
        nicNumber: string;
        passportNumber: string;
        phoneNumber: string;
        email: string;
        visitorId: string;
        visitorIdentityType: string;
        visitorStatus?: string
}